Загрузка файлов на сервер работает корректно.

+++ cloud-storage-server/src/main/java/com/vorobyov/cloudstorage/server/handlers/UploadFileHandler.java
@@ -1,6 +1,5 @@
 package com.vorobyov.cloudstorage.server.handlers;
 
-import com.vorobyov.cloudstorage.server.utils.ByteArray;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
 
@@ -13,51 +12,62 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.logging.Logger;
 
-public class UploadFileHandler extends SimpleChannelInboundHandler<ByteArray> {
+public class UploadFileHandler extends SimpleChannelInboundHandler<ByteBuffer> {
 	Logger logger = Logger.getLogger(this.getClass().getName());
 	
 	private Path fileToWrite;
 	private long fileSize;
 	private long bytesRead;
+	private RandomAccessFile raf = null;
+	private FileChannel fileChannel = null;
+	private boolean isReadInProgress = false;
 	
 	@Override
-	protected void channelRead0(ChannelHandlerContext ctx, ByteArray msg) throws Exception {
+	protected void channelRead0(ChannelHandlerContext ctx, ByteBuffer msg) throws Exception {
 		logger.info("Starting read channel...");
-		
-		if (Files.exists(fileToWrite)) {
-			Files.delete(fileToWrite);
+		if (!isReadInProgress) {
+			if (Files.exists(fileToWrite)) {
+				Files.delete(fileToWrite);
+			}
+			Files.createFile(fileToWrite);
+			isReadInProgress = true;
 		}
-		Files.createFile(fileToWrite);
 		writeBytesToFile(msg, ctx);
 	}
 	
-	private void writeBytesToFile(ByteArray msg, ChannelHandlerContext ctx) {
-		try (RandomAccessFile raf = new RandomAccessFile(fileToWrite.toString(), "rw");
-		     FileChannel fileChannel = raf.getChannel()) {
+	private void writeBytesToFile(ByteBuffer msg, ChannelHandlerContext ctx) {
+		try {
+			if (raf == null) {
+				raf = new RandomAccessFile(fileToWrite.toString(), "rw");
+				logger.info("raf created");
+				if (fileChannel == null) {
+					fileChannel = raf.getChannel();
+					logger.info("fileChannel created");
+				}
+			}
 			
-//			ByteBuffer[] byteBuffers = byteBuf.nioBuffers();
-			ByteBuffer byteBuffer = ByteBuffer.wrap(msg.getData());
+//			ByteBuffer byteBuffer = ByteBuffer.wrap(msg.getData());
 
-//			for (int i = 0; i < byteBuf.nioBufferCount(); i++) {
-//				while (byteBuffers[i].hasRemaining()) {
-//					fileChannel.position(file.length());
-//					fileChannel.write(byteBuffers[i]);
-//				}
+//			for (ByteBuffer b : msg) {
+				bytesRead += fileChannel.write(msg, bytesRead);
 //			}
 			
-			bytesRead += fileChannel.write(byteBuffer, bytesRead);
 			
 			logger.info("read " + bytesRead + " bytes");
 			logger.info("File size " + Files.size(fileToWrite) + " bytes");
 			
-			if (fileSize == bytesRead) {
+			if (fileSize == Files.size(fileToWrite)) {
 				logger.info("/upload complete");
+				
 				fileSize = 0;
 				fileToWrite = null;
 				bytesRead = 0;
+				isReadInProgress = false;
+				
+				msg.clear();
 				ctx.pipeline().get(ByteBufToByteArrayHandler.class).expectCommand();
 //				ctx.pipeline().lastContext().fireChannelRead("/upload complete".getBytes(StandardCharsets.UTF_8));
-				ctx.fireChannelRead(new ByteArray("/upload complete".getBytes(StandardCharsets.UTF_8)));
+				ctx.fireChannelRead(ByteBuffer.wrap("/upload complete".getBytes(StandardCharsets.UTF_8)));
 
 			}
 				//			} else {
@@ -67,10 +77,22 @@ public class UploadFileHandler extends SimpleChannelInboundHandler<ByteArray> {
 		
 		} catch (IOException e) {
 			e.printStackTrace();
+		} finally {
+			try {
+				if (!isReadInProgress) {
+					if (fileChannel != null) {
+						fileChannel.close();
+						fileChannel = null;
+					}
+					if (raf != null) {
+						raf.close();
+						raf = null;
+					}
+				}
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
 		}
-//		} finally {
-//			byteBuf.release();
-//		}
 	}
 	
 	public void setFileToWrite(Path fileToWrite) {
